Crate audiopus

Audiopus is a high level abstraction over the Opus library.

This crate uses TryFrom to prevent the incorrect use of Opus. The API accepts newtypes such as Packet, MutPacket, and MutSignals. The implementation of TryFrom ensures Opus’ restrictions will be kept in mind by checking these on conversion. Without these restrictions, crashes may occur among others, because Opus does not know any types larger than i32 and does not expect empty packets.

Packet, MutPacket, MutSignals implement conversions from &Vec[T] and &[T], they borrow only. The Mut notes when the newtype borrows mutably.

A Packet references an underlying buffer of type &[u8], it cannot be empty and not longer than std::i32::MAX.

Same goes for MutPacket, except the type mutably borrows the buffer thus the length may change after passing it to Opus. Hence the length of this type will be returned as Result.

MutSignals wraps around a generic buffer and represents Opus’ output. E.g. when encoding, Opus will fill the buffer with the encoded data.

Audiopus aims to never panic or crash when interacting with Opus, if either occurs, consider this a bug and please report it on the GitHub!

Re-exports

pub use crate::error::Error;

pub use crate::error::ErrorCode;

pub use crate::error::Result;

pub use audiopus_sys as ffi;

Modules

coder

error

packet

repacketizer

softclip

Structs

MutSignals

A newtype wrapping around a mutable buffer. They represent mutably borrowed arguments that will be filled by Opus. E.g. you pass this to an encode-method and Opus encodes data into the underlying buffer.

Enums

Application

Represents possible application-types for Opus.

Bandwidth

Represents possible bandwidths of an Opus-stream.

Bitrate

Channels

Represents possible audio channels Opus can use.

SampleRate

Represents possible sample rates Opus can use. Values represent Hertz.

Signal

Functions

version

Gets the libopus version string.